Updated graphs
Filter is only working on the circuit graph, but the day filter is not displaying data for some reason. The Energy Usage chart is currently using data that is saved in a txt file and the filter has not been coded for that. Things to do will be fix the filter, make the graph look more like the figma and probaly add a calendar data picker so that you can chose what day to filter by.